+++ packages/iov-bns/types/decode.d.ts
@@ -8,7 +8,7 @@ import { BnsUsernameNft, CashConfiguration, ElectionRule, Electorate, Keyed, Par
  */
 export declare function asIntegerNumber(maybeLong: Long | number | null | undefined): number;
 export declare function ensure<T>(maybe: T | null | undefined, msg?: string): T;
-export declare function decodeUsernameNft(nft: codecImpl.username.IUsernameToken, registryChainId: ChainId): BnsUsernameNft;
+export declare function decodeUsernameNft(nft: codecImpl.username.IToken & Keyed, registryChainId: ChainId): BnsUsernameNft;
 export declare function decodeNonce(sequence: Long | number | null | undefined): Nonce;
 export declare function decodeUserData(userData: codecImpl.sigs.IUserData): {
     readonly nonce: Nonce;
@@ -24,5 +24,5 @@ export declare function decodeParticipants(prefix: "iov" | "tiov", maybeParticip
 export declare function decodeElectorate(prefix: "iov" | "tiov", electorate: codecImpl.gov.IElectorate & Keyed): Electorate;
 export declare function decodeElectionRule(prefix: "iov" | "tiov", rule: codecImpl.gov.IElectionRule & Keyed): ElectionRule;
 export declare function decodeProposal(prefix: "iov" | "tiov", proposal: codecImpl.gov.IProposal & Keyed): Proposal;
-export declare function parseMsg(base: UnsignedTransaction, tx: codecImpl.app.ITx): UnsignedTransaction;
-export declare function parseTx(tx: codecImpl.app.ITx, chainId: ChainId): SignedTransaction;
+export declare function parseMsg(base: UnsignedTransaction, tx: codecImpl.bnsd.ITx): UnsignedTransaction;
+export declare function parseTx(tx: codecImpl.bnsd.ITx, chainId: ChainId): SignedTransaction;