+++ packages/iov-bns/types/encode.d.ts
@@ -6,6 +6,6 @@ export declare function encodePrivkey(privateKey: PrivkeyBundle): codecImpl.cryp
 export declare function encodeAmount(amount: Amount): codecImpl.coin.ICoin;
 export declare function encodeFullSignature(fullSignature: FullSignature): codecImpl.sigs.IStdSignature;
 export declare function encodeParticipants(participants: readonly Participant[]): codecImpl.multisig.IParticipant[];
-export declare function buildMsg(tx: UnsignedTransaction): codecImpl.app.ITx;
-export declare function buildUnsignedTx(tx: UnsignedTransaction): codecImpl.app.ITx;
-export declare function buildSignedTx(tx: SignedTransaction): codecImpl.app.ITx;
+export declare function buildMsg(tx: UnsignedTransaction): codecImpl.bnsd.ITx;
+export declare function buildUnsignedTx(tx: UnsignedTransaction): codecImpl.bnsd.ITx;
+export declare function buildSignedTx(tx: SignedTransaction): codecImpl.bnsd.ITx;