+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Breaking changes
 
 - @iov/bns: `ElectionRule`, `Proposal` and `VersionedId` all use numeric IDs now
   to conform with `Electorate`.
+- @iov/bns: Remove `AddAddressToUsernameTx` and `UpdateTargetsOfUsernameTx` in
+  favour of the new `UpdateTargetsOfUsernameTx`.
+- @iov/bns: Rename `BnsUsernameNft.addresses` to `.targets`.
+- @iov/bns: Rename `RegisterUsernameTx.addresses` to `.targets`.
 
 ## 0.15.1